Patch #SF3010443, from Kaspar Brand:
1) the success/passed/failure functions from Red Hat's init scripts do not
include a newline (they use echo -n). The simplest fix is to add an echo
statement to the respective od_status functions; if this is undesirable for
other LSB-compliant distros, then a somewhat quirky workaround would be
possible by defining the log_xxx aliases accordingly, but I hope we can do
without.

2) only execute the var/lock/subsys stuff on successful start/stop