contrib/systemd: use @runstatedir@ in opendkim.service.in.
Now that the "runstatedir" variable is supported by the build system,
use it in the systemd service file template. The existing PID file
path was under "$@localstatedir@/run", and this commit changes that
string to "@runstatedir@". Unless the user goes out of his way to
specify a custom runstatedir, these should be equivalent: the value of
runstatedir in configure.ac defaults to $localstatedir/run.