contrib/systemd: enable full @VARIABLE@ substitution.
The existing @variable@ substitutions in contrib/systemd don't quite
work, and the autoconf documentation recommends against the current
approach for the same reason: unless you specify a full, absolute,
path for (say) @sbindir@, then autoconf will only expand the default
value of that variable one level, to "${prefix}/sbin". That's not
usable out of the box, and we have no good way to replace the
"${prefix}" variable with its configure-time value.

This commit switches the substitution tactic to perform the
replacement in the Makefile, rather than with autoconf. This is in
fact what the autoconf documentation recommends, even if it looks
silly. Instead of (say) @sbindir@, we use a custom @SBINDIR@ variable,
and then set it to the appropriate value with sed at build-time. The
value of the variable itself still comes from the build system.