opendkim/opendkim.conf.*.in: use @runstatedir@ in PidFile paths.
Now that the @runstatedir@ variable is available, use it in the
(commented-out) PidFile settings present in the two sample
configuration files.

The example location for these two files has also been moved out of
the "opendkim" subdirectory of @runstatedir@ for security reasons.
Generally, that subdirectory will be writable by a restricted user as
whom the daemon is running, so that, for example, a local socket can
be created there. The PID file, on the other hand, needs to be
writable only by root, because root will generally be the one who
calls "kill" on its contents. OpenDKIM handles this situation
correctly: the PID file is created as root, even when dropping
privileges, so the PID file is successfully written directly to
@runstatedir@.