opendkim/opendkim.conf.*.in: enable full @VARIABLE@ substitution.
The existing @runstatedir@ substitutions in the example configuration
files don't quite work, and the autoconf documentation recommends
against the current approach for the same reason: unless you specify a
full, absolute path for @runstatedir@, then autoconf will only expand
the default value of that variable one level, to "${localstatedir}/run".
That's not usable out of the box, and we have no good way to replace
the "${localstatedir}" variable with its configure-time value.

This commit switches the substitution tactic to perform the
replacement in the Makefile, rather than with autoconf. This is in
fact what the autoconf documentation recommends, even if it looks
silly. Instead of @runstatedir@, we use a custom @RUNSTATEDIR@
variable, and then set it to the appropriate value with sed at
build-time. The value of the variable itself still comes from the
build system.