autotools: refactor build-time @VARIABLE@ substitutions.
Some recent commits introduced build-time variable substitution in
three places,

  1. contrib/OpenRC/*
  2. contrib/systemd/*
  3. opendkim/opendkim.conf.*.in

Each of these uses essentially the same strategy, making it a good
candidate to be factored out. This commit does just that, moving
everything into the top-level Makefile.am, which is the closest parent
of both the "contrib" and "opendkim" directories.

Ultimately, the behavior of the rules is unchanged.