Add implementations of strlcpy and strlcat to opendkim
Implementations were taken from libbsd to maintain compatibility.

List of changes and their justification:

configure
- rebuild configure script because of changes in configure.ac

configure.ac
- added subdir-objects to automake options because linux_string folder
is used in multiple makefiles and is in subdirectory (some warnings are
displayed without it)
- Removed check for getgroups function because
1. it is no longer used in any source file
2. in my opinion bug in autoconf was forcing configure script to link
libbsd to get getgroups function even if it is included in glibc
- added template, argument and automake conditional for --use-bundled-strl
argument. It was neccessary to change structure of some conditionals
because of this change.

libopendkim/Makefile.am
- added automake conditionals for USE_BUNDLED_STRL

opendkim/Makefile.am
- added automake conditionals for USE_BUNDLED_STRL

*.c files
- added build time conditionals for USE_BUNDLED_STRL macro