configure.ac: drop lua5.1 pkg-config check.
The official upstream name for the lua pkg-config file is "lua.pc", so
preferring the name "lua5.1.pc" to the official one can lead to
problems when both are installed. Specifically, if lua-5.2 is
installed and its pkg-config file is called "lua.pc", then ./configure
will still detect (and use!) lua-5.1 if it happens to be installed
with just the right name.

This was honestly left in only for superstitious reasons, and now it's
actually causing problems on Gentoo systems that have more than one
version of Lua installed. So, time to go.

Closes: https://github.com/trusteddomainproject/OpenDKIM/issues/111