autotools: replace Lua detection madness.
This commit dismantles the existing lua-detecting machine in
configure.ac and replaces it with something much simpler. First, the
--with-lua flag has been replaced by an --enable-lua flag that takes
no arguments other than the implicit yes/no. If --enable-lua is
passed, you are expected to have Lua somewhere on your machine in a
place where your toolchain can find it. That can be accomplished in a
few ways:

  1. by using pkg-config; or

  2. by installing one copy of Lua with the standard names in the
     usual system locations; or

  3. by telling your toolchain where to look for headers and libraries
     manually with the -I, -l, and -L flags.

This should cover 100% of users. If we can find lua.pc (the official
upstream name) in your PKG_CONFIG_PATH, it will be used. If not, we
will also (as a courtesy to Debian users) try the non-standard
lua5.1.pc and lua5.2.pc names. If those are found, they will be used.
If pkg-config cannot be used for some reason, then as a fallback, we
search for liblua on the system using AC_SEARCH_LIBS. If you have Lua
installed on your system with the standard names, this will find it.

The only remaining scenario is when you have Lua installed to a
nonstandard location and pkg-config is unavailble. Even this can be
handled with the standard tooling; for example, if your headers are
installed to /usr/include/lua5.2 and if your libraries are in
/usr/lib64/lua5.2, then you can set

  CPPFLAGS="I/usr/include/lua5.2" LIBS="-L/usr/lib64/lua5.2"

in your environment during the build, and that copy of Lua will be
found and used just as if those two paths were part of the system
defaults. If, in addition, someone has perversely renamed your library
to something like lua5.2.so, then "-llua5.2" can also be appended to
LIBS; although beware that your executables may be overlinked as a
result. (Overlinking is generally harmless.)