configure.ac: update OpenLDAP's pkgconfig name
OpenLDAP provides the file ldap.pc for its libldap library. This can
be verified via libraries/libldap/ldap.pc.in in the repository,

  https://git.openldap.org/openldap/openldap/-/blob/master/

Our ./configure script checks instead for the name "openldap", which
at some point may have been correct, but no longer works. We switch to
"ldap" so that we can locate the upstream file.

On some platforms (https://bugs.gentoo.org/918512) this will fix a
linking error. Thanks to Chris Pritchard for the report and the
diagnosis.