configure.ac: update Cyrus SASL's pkgconfig name
Cyrus SASL provides the file libsasl2.pc for its libsasl2
library. This can be verified in its git repository:

  https://github.com/cyrusimap/cyrus-sasl/blob/master/libsasl2.pc.in

Our ./configure script checks instead for the name "cyrussasl", which
at some point may have been correct, but no longer works. We switch to
"libsasl2" so that we can locate the upstream file.

On some platforms (https://bugs.gentoo.org/918512) this will fix a
linking error. Thanks to Chris Pritchard for the report and the
diagnosis.